+++ node/components/network/src/proto/gossip.proto
@@ -20,9 +20,14 @@ message PushValidatorAddrs {
   repeated roles.validator.Signed net_addresses = 1;
 }
 
-message PushTx {
+message Transaction {
   // Transaction bytes.
   optional bytes tx = 1; // required
+}
+
+message PushTx {
+  // Transaction
+  optional Transaction tx = 1; // required
 }
 
 message Last {
