+++ node/libs/engine/Cargo.toml
@@ -11,10 +11,12 @@ version.workspace    = true
 
 [dependencies]
 zksync_concurrency.workspace      = true
+zksync_consensus_crypto.workspace = true
 zksync_consensus_roles.workspace  = true
 
 anyhow.workspace      = true
 async-trait.workspace = true
+hex.workspace         = true
 rand.workspace        = true
 tracing.workspace     = true
 vise.workspace        = true
