+++ node/libs/roles/src/node/tests.rs
@@ -17,9 +17,6 @@ fn test_byte_encoding() {
 
 #[test]
 fn test_text_encoding() {
-    let ctx = &ctx::test_root(&ctx::RealClock);
-    let rng = &mut ctx.rng();
-
     let key = SecretKey::generate();
     let t1 = TextFmt::encode(&key);
     let t2 = TextFmt::encode(&key.public());
@@ -27,10 +24,6 @@ fn test_text_encoding() {
     assert_eq!(key.public(), Text::new(&t2).decode().unwrap());
     assert!(Text::new(&t1).decode::<PublicKey>().is_err());
     assert!(Text::new(&t2).decode::<SecretKey>().is_err());
-
-    let tx: TxHash = rng.gen();
-    let t1 = TextFmt::encode(&tx);
-    assert_eq!(tx, Text::new(&t1).decode::<TxHash>().unwrap());
 }
 
 #[test]
